runner.rb: make sure entire output is written to socket
The previous implementation would read and write a 1-byte buffer from
the build.sh script to the socket, but the thread doing this is
immediately closed once the process is finished. This makes it possible
for some output not to have been read and written to the socket before
the thread is closed, and we could lose some output at the end of a
command.

The new implementation uses IO.copy_stream instead, which will copy the
output from one IO object into another, and will block until the source
object reaches an EOF. The current way of spawning the process with
PTY.open and Process.spawn causes the "output" part of the pipe (the
`io` variable) to not be closed at the end of the command, so that is
changed to PTY.spawn, which does close the output IO object.