Use systemsetup to automate some settings
This enforces the system sleep settings, as well as enabling remote
login and setting the time zone. The relevant things have been removed
from the README, since they now will be handled by the bootstrap script
instead, with the exception of selecting the location, since that's a
required step in the setup.